No warnings

+++ cpp/ycm/ClangCompleter/ClangHelpers.cpp
@@ -50,7 +50,7 @@ DiagnosticKind DiagnosticSeverityToType( CXDiagnosticSeverity severity ) {
 FixIt BuildDiagnosticFixIt( const std::string& text, CXDiagnostic diagnostic ) {
   FixIt fixit;
 
-  size_t num_chunks = clang_getDiagnosticNumFixIts( diagnostic );
+  unsigned int num_chunks = clang_getDiagnosticNumFixIts( diagnostic );
   if ( !num_chunks ) {
     return fixit;
   }
@@ -59,7 +59,7 @@ FixIt BuildDiagnosticFixIt( const std::string& text, CXDiagnostic diagnostic ) {
   fixit.location = Location( clang_getDiagnosticLocation( diagnostic ) );
   fixit.text = text;
 
-  for ( size_t idx = 0; idx < num_chunks; ++idx ) {
+  for ( unsigned int idx = 0; idx < num_chunks; ++idx ) {
     FixItChunk chunk;
     CXSourceRange range;
     chunk.replacement_text = CXStringToString(
@@ -107,13 +107,13 @@ void BuildFullDiagnosticDataFromChildren(
     return;
   }
 
-  size_t num_child_diagnostics = clang_getNumDiagnosticsInSet( diag_set );
+  unsigned int num_child_diagnostics = clang_getNumDiagnosticsInSet( diag_set );
 
   if ( !num_child_diagnostics ) {
     return;
   }
 
-  for ( size_t i = 0; i < num_child_diagnostics; ++i ) {
+  for ( unsigned int i = 0; i < num_child_diagnostics; ++i ) {
     CXDiagnostic child_diag = clang_getDiagnosticInSet( diag_set, i );
 
     if( !child_diag ) {
@@ -140,10 +140,11 @@ bool CompletionStringAvailable( CXCompletionString completion_string ) {
 
 std::vector< Range > GetRanges( const DiagnosticWrap &diagnostic_wrap ) {
   std::vector< Range > ranges;
-  size_t num_ranges = clang_getDiagnosticNumRanges( diagnostic_wrap.get() );
+  unsigned int num_ranges =
+      clang_getDiagnosticNumRanges( diagnostic_wrap.get() );
   ranges.reserve( num_ranges );
 
-  for ( size_t i = 0; i < num_ranges; ++i ) {
+  for ( unsigned int i = 0; i < num_ranges; ++i ) {
     ranges.emplace_back( clang_getDiagnosticRange( diagnostic_wrap.get(), i ) );
   }
 
@@ -215,7 +216,7 @@ std::vector< CompletionData > ToCompletionDataVector(
   completions.reserve( results->NumResults );
   unordered_map< std::string, size_t > seen_data;
 
-  for ( size_t i = 0; i < results->NumResults; ++i ) {
+  for ( unsigned int i = 0; i < results->NumResults; ++i ) {
     CXCompletionResult result = results->Results[ i ];
     CXCompletionString completion_string = result.CompletionString;
